Use react query
